Added some error handling
Check if file exists, is a PDF and if a watermark was given.
Also only create .temp if there is no such directory yet.

+++ unwatermarkpdf.sh
@@ -7,12 +7,21 @@
 # Uncomment last line for destructive editing.
 #
 range=${3-1-end}
-mkdir .temp
-pdftk $1 cat $range output .temp/uncompressed.pdf uncompress
-cd .temp
-sed -e "s/$2/ /g" <uncompressed.pdf >unwatermarked.pdf
-pdftk unwatermarked.pdf output fixed.pdf
-pdftk fixed.pdf output ../clean_$1 compress
-cd ..
-rm -r .temp
-# mv clean_$1 $1
+
+if [ ! -f $1 ]; then
+	echo "File does not exist."
+elif [ ! ${1: -4} == ".pdf" ]; then
+	echo "This is not a PDF"
+elif [ -z ${2+x} ]; then
+	echo "No watermark given"
+else
+	mkdir -p .temp
+	pdftk $1 cat $range output .temp/uncompressed.pdf uncompress
+	cd .temp
+	sed -e "s/$2/ /g" <uncompressed.pdf >unwatermarked.pdf
+	pdftk unwatermarked.pdf output fixed.pdf
+	pdftk fixed.pdf output ../clean_$1 compress
+	cd ..
+	rm -r .temp
+	# mv clean_$1 $1
+fi
